requirements.txt
langchain
langchain-community
langchain-chroma
langchain-text-splitters
langgraph
python-dotenv
pydantic
tiktoken
beautifulsoup4
pypdf